Joined test_collusion_attacks.py and test_local_attacks.py to test_attacks.py
