+++ tests/test_cases/test_attacks.py
@@ -16,11 +16,23 @@ import unittest
 
 import numpy as np
 
+from shared.attacks.collusion import DeltaToZeroAttack, ScalarPAFAttack
 from shared.attacks.local import ConstantAttack, SignFlipAttack
 
 
 def gen_rand_vec(dn=10):
   return np.random.randn(dn)
+
+
+class CollusionAttacksTest(unittest.TestCase):
+  def test_delta_to_zero(self):
+    v = gen_rand_vec()
+    np.testing.assert_array_almost_equal(-v, DeltaToZeroAttack()(v))
+
+  def test_paf(self):
+    const = 100
+    v = gen_rand_vec()
+    np.testing.assert_array_almost_equal(v + const, ScalarPAFAttack(const)(v))
 
 
 class LocalAttacksTest(unittest.TestCase):
